refactor image info
